AB#27292 merge from dev updates
